+++ service-kit-core/src/settings/configuration.rs
@@ -1,0 +1,36 @@
+use serde::Deserialize;
+use std::path::PathBuf;
+
+#[derive(Clone, Debug, Deserialize)]
+pub struct Configuration {
+    pub db: Option<Database>,
+    pub storage: Option<Storage>,
+}
+
+#[derive(Clone, Debug, Deserialize)]
+pub struct Storage {
+    pub path: PathBuf,
+}
+
+#[derive(Clone, Debug, Deserialize)]
+pub struct Database {
+    host: String,
+    port: u16,
+    user: String,
+    password: String,
+    database: Option<String>,
+}
+
+impl std::fmt::Display for Database {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "mysql://{}:{}@{}:{}/{}",
+            self.user,
+            self.password,
+            self.host,
+            self.port,
+            self.database.as_deref().unwrap_or("acore_world"),
+        )
+    }
+}
